+++ examples/commercial-quick-order/api-types/storeApiTypes.d.ts
@@ -2,7 +2,7 @@
  * This file is auto-generated. Do not make direct changes to the file.
  * Instead override it in your shopware.d.ts file.
  *
- * Shopware API version: unknown
+ * Shopware API version: 6.6.8.0
  *
  */
 type GenericRecord =
@@ -859,7 +859,7 @@ export type Schemas = {
     messageKey: string;
   };
   CartItems: {
-    items?: components["schemas"]["LineItem"][];
+    items: components["schemas"]["LineItem"][];
   };
   CartListPrice: {
     /** @enum {string} */
@@ -4752,8 +4752,7 @@ export type Schemas = {
       label: string;
       priority: number;
       translated: {
-        apiAlias?: string;
-        key?: string;
+        key: string;
         label: string;
       };
     }[];
@@ -4768,8 +4767,8 @@ export type Schemas = {
         min: number;
       };
       properties: string[];
-      rating?: number; // TODO: [OpenAPI][ProductListingResult] - rating should be defined the same as in body of the request
-      search: string; // TODO: [OpenAPI][ProductListingResult] - search should be required as is required in body of the request, otherwise everywhere optional
+      rating: number | null;
+      search?: string;
       /** @default false */
       "shipping-free": boolean;
     };
@@ -7401,18 +7400,8 @@ export type Schemas = {
     readonly updatedAt?: string;
   };
   SwagPaypalVaultToken: {
-    // TODO: [OpenAPI][SwagPaypalVaultToken] - add SwagPaypalVaultToken definition to schema
-    /** Format: date-time */
-    createdAt: string;
-    customer?: components["schemas"]["Customer"];
-    customerId: string;
     id?: string;
-    identifier: string;
-    mainMapping?: components["schemas"]["SwagPaypalVaultTokenMapping"];
-    paymentMethod?: components["schemas"]["PaymentMethod"];
-    paymentMethodId: string;
-    /** Format: date-time */
-    updatedAt?: string;
+    identifier?: string;
   };
   SwagPaypalVaultTokenMapping: {
     /** Format: date-time */
@@ -7982,7 +7971,7 @@ export type operations = {
       /** Customer last name. Value will be reused for shipping and billing address if not provided explicitly. */
       lastName: string;
       /** Id of the salutation for the customer account. Fetch options using `salutation` endpoint. */
-      salutationId: string;
+      salutationId?: string;
       /** (Academic) title of the customer */
       title?: string;
     } & (
@@ -8175,7 +8164,7 @@ export type operations = {
       /** Password for the customer. Required, unless `guest` is `true` */
       password: string;
       /** Id of the salutation for the customer account. Fetch options using `salutation` endpoint. */
-      salutationId: string;
+      salutationId?: string;
       shippingAddress?: components["schemas"]["CustomerAddress"];
       /** URL of the storefront for that registration. Used in confirmation emails. Has to be one of the configured domains of the sales channel. */
       storefrontUrl: string;
@@ -8470,11 +8459,16 @@ export type operations = {
       "sw-language-id"?: string;
     };
     body: {
-      // TODO: [OpenAPI][updateLineItem] - add proper request body type with required fields
-      items: Array<{
-        id: string;
-        quantity: number;
-      }>;
+      items: [
+        {
+          id: string;
+          quantity: number;
+        },
+        ...{
+          id: string;
+          quantity: number;
+        }[],
+      ];
     };
     response: components["schemas"]["Cart"];
     responseCode: 200;
@@ -8584,7 +8578,7 @@ export type operations = {
       /** Phone. This field may be required depending on the system settings. */
       phone?: string;
       /** Identifier of the salutation. Use `/api/salutation` endpoint to fetch possible values. */
-      salutationId: string;
+      salutationId?: string;
       /** Identifier of the cms element */
       slotId?: string;
       /** The subject of the contact form. */
@@ -8991,6 +8985,16 @@ export type operations = {
     } & components["schemas"]["EntitySearchResult"];
     responseCode: 200;
   };
+  "readB2bEmployee get /employee/{id}": {
+    contentType?: "application/json";
+    accept?: "application/json";
+    pathParams: {
+      /** Identifier of the employee to be read */
+      id: string;
+    };
+    response: components["schemas"]["B2bEmployee"];
+    responseCode: 200;
+  };
   "readEmployee post /employee/{id}": {
     contentType?: "application/json";
     accept?: "application/json";
@@ -9383,7 +9387,7 @@ export type operations = {
     };
     body?: components["schemas"]["Criteria"];
     response: {
-      elements: components["schemas"]["Product"][]; // TODO: [OpenAPI][readProduct]: add elements property as required
+      elements: components["schemas"]["Product"][];
     } & components["schemas"]["EntitySearchResult"];
     responseCode: 200;
   };
@@ -9883,11 +9887,11 @@ export type operations = {
     body?: components["schemas"]["Criteria"];
     response: {
       /** aggregation result */
-      aggregations?: Record<string, never>;
-      elements: components["schemas"]["ShippingMethod"][]; // TODO: [OpenAPI][readShippingMethod]: response should be `EntitySearchResult` and elements should be required
+      aggregations?: GenericRecord;
+      elements: components["schemas"]["ShippingMethod"][];
       /** Total amount */
       total?: number;
-    };
+    } & components["schemas"]["EntitySearchResult"];
     responseCode: 200;
   };
   "createShoppingList post /shopping-list": {