+++ examples/snippets-middleware/components/Frontends.vue
@@ -1,7 +1,50 @@
 <script setup lang="ts">
+import type { Schemas } from "#shopware";
+
 const { currency, selectedPaymentMethod } = useSessionContext();
+const { getLanguageIdFromCode, getAvailableLanguages } =
+  useInternationalization();
+
 const { isLoggedIn } = useUser();
 const { t, availableLocales, locale: currentLocale } = useI18n();
+const { resolvePath } = useNavigationSearch();
+const { apiClient } = useShopwareContext();
+const backendRoute = ref<Schemas["SeoUrl"] | null>();
+
+// function to fetch the pretty URL for a different language
+const findRouteForLanguage = async (languageCode: string) => {
+  if (!backendRoute.value?.foreignKey) {
+    return;
+  }
+  const languageId = await getLanguageIdFromCode(languageCode);
+  const seoResult = await apiClient.invoke("readSeoUrl post /seo-url", {
+    headers: {
+      "sw-language-id": languageId,
+    },
+    body: {
+      filter: [
+        {
+          type: "equals",
+          field: "foreignKey",
+          value: backendRoute.value?.foreignKey,
+        },
+      ],
+    },
+  });
+  return seoResult.data.elements?.[0];
+};
+
+watch(currentLocale, async () => {
+  // fetch the new pretty URL for a different language
+  backendRoute.value = await findRouteForLanguage(unref(currentLocale));
+});
+
+onMounted(async () => {
+  // fetch the available languages
+  await getAvailableLanguages();
+  const response = await resolvePath("/Summer-trends/");
+  backendRoute.value = response;
+});
 </script>
 
 <template>
@@ -25,6 +68,10 @@ const { t, availableLocales, locale: currentLocale } = useI18n();
       </a>
     </div>
     <p>
+     {{ t("general.url") }}:
+      <strong>{{ backendRoute?.seoPathInfo }}</strong>
+    </p>
+    <p>
       {{ t("general.currency") }}:
       <strong>{{ currency?.name }} ({{ currency?.symbol }})</strong>
     </p>
@@ -38,9 +85,9 @@ const { t, availableLocales, locale: currentLocale } = useI18n();
     </p>
   </div>
   <div>
-    you can also examine what is returned by visiting
-    <strong>/api/translations?locale={{ currentLocale }}</strong> and loaded by
-    i18n module
+    you can also examine what is returned from an API middleware and loaded by
+    i18n module by visiting
+    <strong><a :href="`${useRequestURL()}api/translations?locale=${currentLocale}`" target="_blank">{{ useRequestURL()}}api/translations?locale={{currentLocale}}</a></strong> 
   </div>
 </template>
 <style scoped>