+++ examples/use-cart/app.vue
@@ -46,11 +46,11 @@ onMounted(async () => {
       class="w-full max-w-sm bg-white rounded-lg shadow-md dark:bg-gray-800 dark:border-gray-700"
     >
       <div class="px-5 pb-5">
-        <h5
+        <div
           class="text-xl font-semibold tracking-tight text-gray-900 dark:text-white"
         >
           {{ product?.translated.name }}
-        </h5>
+        </div>
 
         <div class="flex items-center justify-between mt-4">
           <span class="text-3xl font-bold text-gray-900 dark:text-white">{{