+++ packages/api-gen/src/commands/generate.ts
@@ -17,17 +17,36 @@ import {
   displayPatchingSummary,
   loadApiGenConfig,
   loadJsonOverrides,
+  resolveSinglePath,
 } from "../jsonOverrideUtils";
 import { extendedDefu, patchJsonSchema } from "../patchJsonSchema";
 import { processAstSchemaAndOverrides } from "../processAstSchemaAndOverrides";
 import { transformOpenApiTypes } from "../transformOpenApiTypes";
 import { transformSchemaTypes } from "../transformSchemaTypes";
 
+/**
+ * Generate schema from your API instance
+ */
 export async function generate(args: {
+  /**
+   * Current working directory
+   */
   cwd: string;
+  /**
+   * Filename of the schema to process, default is `storeApiSchema.json` or `adminApiSchema.json` depending on the `apiType` parameter
+   */
   filename?: string;
+  /**
+   * Type of the API to generate types for
+   */
   apiType: "store" | "admin";
+  /**
+   * Debug mode, display additional information and generates additional files, not needed for the regular usage
+   */
   debug: boolean;
+  /**
+   * Log patches, display information about applied patches while generating types
+   */
   logPatches: boolean;
 }) {
   const inputFilename = args.filename
@@ -41,10 +60,9 @@ export async function generate(args: {
     const fullInputFilePath = join(args.cwd, "api-types", inputFilename);
     const fullOutputFilePath = join(args.cwd, "api-types", outputFilename);
 
-    //check if file exist
-    const fileExist = existsSync(fullInputFilePath);
+    const resolvedSchema = await resolveSinglePath<OpenAPI3>(inputFilename);
 
-    if (!fileExist && !args.apiType) {
+    if (!resolvedSchema && !args.apiType) {
       console.log(
         c.yellow(
           `Schema file ${c.bold(
@@ -61,12 +79,9 @@ export async function generate(args: {
     let processedSchemaAst: TransformedElements;
     let apiVersion = "unknown";
 
-    if (fileExist) {
+    if (resolvedSchema) {
       // Apply patches
-      const schemaFile = readFileSync(fullInputFilePath, {
-        encoding: "utf-8",
-      });
-      const schemaForPatching = json5.parse(schemaFile) as OpenAPI3;
+      const schemaForPatching = structuredClone(resolvedSchema);
       apiVersion = schemaForPatching?.info?.version;
 
       const configJSON = await loadApiGenConfig({
@@ -103,7 +118,7 @@ export async function generate(args: {
         jsonOverrides,
       });
 
-      const originalSchema = json5.parse(schemaFile);
+      const originalSchema = structuredClone(resolvedSchema);
       console.log("schema", originalSchema.info);
 
       displayPatchingSummary({