+++ packages/composables/src/useListing/useListing.ts
@@ -525,7 +525,9 @@ export function createListingComposable({
       };
     }
 
-    return search(appliedFilters);
+    return search(
+      appliedFilters as operations["searchPage post /search"]["body"],
+    );
   };
 
   const resetFilters = () => {