+++ templates/vue-demo-store/pages/checkout/index.vue
@@ -168,9 +168,6 @@ const termsBox = useTemplateRef("termsBox");
 const shippingMethodBox = useTemplateRef("shippingMethodBox");
 
 const rules = computed(() => ({
-  salutationId: {
-    required,
-  },
   firstName: {
     required,
     minLength: minLength(3),
@@ -328,15 +325,15 @@ const beforeCreateOrderValidation = () => {
 const { updatePersonalInfo } = useUser();
 
 const handleChangeBaseInfo = async (data: {
-  firstName: string;
-  lastName: string;
-  salutationId: string;
+  firstName?: string;
+  lastName?: string;
+  salutationId?: string;
 }) => {
   try {
     await updatePersonalInfo({
-      firstName: data.firstName,
-      lastName: data.lastName,
-      salutationId: data.salutationId,
+      firstName: data.firstName ?? "",
+      lastName: data.lastName ?? "",
+      salutationId: data.salutationId ?? "",
     });
   } catch (error) {
     if (error instanceof ApiClientError)
@@ -441,7 +438,6 @@ const handleChangeBaseInfo = async (data: {
                     autocomplete="on"
                     class="mt-1 block w-full p-2.5 border border-secondary-300 text-secondary-900 text-sm rounded-md shadow-sm focus:ring-brand-light focus:border-light"
                     data-testid="checkout-pi-salutation-select"
-                    @blur="$v.salutationId.$touch()"
                   >
                     <option disabled selected value="">
                       {{ $t("form.chooseSalutation") }}
@@ -454,12 +450,6 @@ const handleChangeBaseInfo = async (data: {
                       {{ salutation.displayName }}
                     </option>
                   </select>
-                  <span
-                    v-if="$v.salutationId.$error"
-                    class="pt-1 text-sm text-red-600 focus:ring-primary border-secondary-300"
-                  >
-                    {{ $v.salutationId.$errors[0].$message }}
-                  </span>
                 </div>
                 <div class="col-span-6 sm:col-span-3">
                   <label